Split out logOnly method from log
Previously log would always call .exit(), which is intended to be fatal.
In order to support non-fatal logging, split out a logOnly method which
outputs the formatted lines but does not exit.  This preserves the
current API completely.

Also add a call to .slice() if the lines argument is not a string so
that the argument Array is not modified.  This codepath is not currently
used, but it is no longer a waiting surprise for future callers.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>