solves the windows issue (with symlinks) (#84)
* solved the windows issue with symlinks

Instead of symlinking to the hook in node-modules/pre-commit, this injects a new pre-commit hook (in .git/hooks) that takes care of launching the hook in node-modules/pre-commit. 
In addition it stashes the unstaged changes and applies them again after running the pre-commit hook, so that you only run your pre-commit scripts on the commit actual codebase.

* removed stashing

Stashing automatically is actually a bad idea, it's better to for the user to have control over doing it or not (manually then).